+++ include/WaveManager.hpp
@@ -46,8 +46,6 @@ public:
    */
   WaveManager(shared_ptr<EventManager> eventManager, shared_ptr<TextLoader> textLoader);
 
-  void delegateMethod(const EventInterface& event);
-
   /*
    * Setup the possible enemies that can spawn
    */