+++ src/GameLogic.cpp
@@ -19,6 +19,7 @@ GameLogic::GameLogic(shared_ptr<TextLoader> textLoader, int windowX, int windowY
   this -> waveManager = unique_ptr<WaveManager>(new WaveManager(eventManager, textLoader));
   this -> projectileManager = unique_ptr<ProjectileManager>(new ProjectileManager(eventManager));
   this -> registerDelegates();
+
   this -> windowX = windowX;
   this -> windowY = windowY;
 }
@@ -36,7 +37,6 @@ void GameLogic::registerDelegates(){
   EventType keyPressEventType = keyPressEvent.getEventType();
   //register the delegate and its type
   this -> eventManager -> registerDelegate(keyPressDelegate, textLoader -> getString(string("IDS_GLD_KP")),keyPressEventType);
-
   //bind our delegate function for mouse presses
   EventManager::EventDelegate mousePressDelegate = std::bind(&GameLogic::handleMousePress, this, _1);